% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ABratio.r
\name{calculate_ABratio}
\alias{calculate_ABratio}
\title{calculate_ABratio}
\usage{
calculate_ABratio(df_rna, df_dna, z, ABminCut, ABminCut_noCI, Threshold)
}
\arguments{
\item{df_rna}{rna allele counts}

\item{df_dna}{dna allele counts}
}
\value{
abratio and confidence interval and the result of genes being balance/imbalance
}
\description{
Function to compute AI ratio and confidence intervals
}
\examples{
calculate_ABratio(df_rna, df_dna, z, ABminCut, ABminCut_noCI)
}
