% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_read_counts.r
\name{df_read_counts}
\alias{df_read_counts}
\title{df_read_counts}
\usage{
df_read_counts(file_path)
}
\arguments{
\item{file_path}{path to the file}
}
\value{
df contains allele counts
}
\description{
Function to read allelic counts
}
\examples{
df_read_counts(file_path) 
}
