% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_samples.r
\name{process_samples}
\alias{process_samples}
\title{process_samples}
\usage{
process_samples(
  samples,
  input_rna,
  input_dna,
  output_dir,
  z = 1.96,
  ABminCut = 1,
  ABminCut_noCI = 0.6,
  Threshold = 10
)
}
\arguments{
\item{samples}{Description of the first parameter}

\item{input_rna}{input RNA df}

\item{input_dna}{input DNA df}

\item{output_dir}{output path}
}
\value{
the output print of IB analysis
}
\description{
Main function to process multiple samples
}
\examples{
process_samples(samples, input_rna, input_dna, output_dir, z=1.96, ABminCut=1, ABminCut_noCI=0.6)
}
